Plane, Visual & geometry updates
- `Visual`
  - manually impl `Debug`.
  - move `VisualGeometry` skeleton to `geometry` submodule.
  - change `VisualBuilder` method `transparency` to accept `Option<impl Into<Rgba>>`.
- `Plane`
  - add methods: `putrstr_at`, `putstr_at_yx`, `putrstr_stained`, `putstr_aligned`, `putstr_aligned_stained`, `putstr_len`, `putstr_len_at`, `putstr_len_at_yx`, `putstr_len_aligned`,
  - make `putstr*` methods to return `u32`.
  - refactor method `resize`.
  - move `PlaneGeometry` to `geometry` submodule.
- move tuples module to the root.
- add `Size` getters `y`, `x`.
re
- update macros, examples and docs.

+++ examples/plane.rs
@@ -48,7 +48,7 @@ fn main() -> Result<()> {
     let mut p1 = Plane::new_sized(&mut nc, (4, 4))?;
     p1.set_scrolling(true);
 
-    assert_eq!["hello world".len(), p1.putstr("hello world")?];
+    assert_eq!["hello world".len() as u32, p1.putstr("hello world")?];
 
     p1.render()?;
